LayoutTests/http/tests/security/cross-origin-css-7.html

<!DOCTYPE html>
<html>
<head>
<title>Cross-origin CSS</title>

<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>

<style>
@import "resources/xorigincss5.html";
</style>
<script>
function getBackgroundColorForId(id) {
    return window.getComputedStyle(document.getElementById(id), null).getPropertyValue('background-color')
}

var onloadTest = async_test("Testing cross-origin and MIME behavior for CSS.");

window.onload = function () {
    test(function () {
        assert_equals(getBackgroundColorForId('id7'), 'rgba(0, 0, 0, 0)');
    }, 'xorigincss5.html should not be loaded.');

    onloadTest.done();
};
</script>
</head>
<body>
    <div id="id7" class="id7"></div>
</body>
</html>
